% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{app,SASEGConnection-method}
\alias{app,SASEGConnection-method}
\title{Access to the SASEGApplication object stored in a SASEGConnection object}
\usage{
\S4method{app}{SASEGConnection}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{SASEGConnection}}.}

\item{...}{Other parameters passed to method. Not used.}
}
\value{
A \code{\linkS4class{SASEGApplication}} object.
}
\description{
\code{app()} method access to the \code{\linkS4class{SASEGApplication}} 
    object stored in a \code{\linkS4class{SASEGConnection}} object. This 
    method is not exported. Only developpers may need to use it.
}
\keyword{internal}
