% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{app,SASEGDriver-method}
\alias{app,SASEGDriver-method}
\title{Access to slot app of a SASEGDriver object}
\usage{
\S4method{app}{SASEGDriver}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{SASEGDriver}}.}

\item{...}{Other parameters passed to method. Not used.}
}
\description{
\code{app()} access to the slot \code{app} of a \code{\linkS4class{SASEGDriver}}.
}
\keyword{internal}
