% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{app}
\alias{app}
\title{Access to slot app of an object}
\usage{
app(obj, ...)
}
\arguments{
\item{obj}{An object with a slot named \code{app}.}

\item{...}{Other parameters passed to method.}
}
\description{
Method to access a slot named \code{app}.
}
\seealso{
\code{\link{app,SASEGDriver-method}}, 
    \code{\link{app,SASEGConnection-method}}
}
\keyword{internal}
