% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{clrCall}
\alias{clrCall}
\title{Call a method on an object (generic)}
\usage{
clrCall(obj, methodName, ...)
}
\arguments{
\item{obj}{An object.}

\item{methodName}{A character string with the name of a method of the object.}

\item{...}{Additional method arguments.}
}
\description{
This is a generic version of the \code{\link[rClr]{clrCall}} function of the \code{rClr} package.
}
\seealso{
\code{cobjRef} method: \code{\link[rClr]{clrCall}}

\code{SASObjRef} method: \code{\link[=clrCall,SASObjRef-method]{clrCall}}
}
\keyword{internal}
