% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{clrGet}
\alias{clrGet}
\title{Get the value of a field or property of an object or class (generic)}
\usage{
clrGet(objOrType, name)
}
\arguments{
\item{objOrType}{An object, or type name, possibly namespace and assembly 
qualified type name.}

\item{name}{The name of a field/property of the object.}
}
\description{
This is a generic version of the \code{\link[rClr]{clrGet}} function of the \code{rClr} 
package.
}
\seealso{
\code{cobjRef} method: \code{\link[rClr]{clrGet}}

\code{SASObjRef} method: \code{\link[=clrGet,SASObjRef-method]{clrGet}}
}
\keyword{internal}
