% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{clrSet}
\alias{clrSet}
\title{Set the value of a field or property of an object or class (generic)}
\usage{
clrSet(objOrType, name, value)
}
\arguments{
\item{objOrType}{An object, or type name, possibly namespace and assembly 
qualified type name.}

\item{name}{The name of a field/property of the object.}

\item{value}{The value to set the field with.}
}
\description{
This is a generic version of the \code{\link[rClr]{clrSet}} function of the 
\code{rClr} package.
}
\seealso{
\code{cobjRef} method: \code{\link[rClr]{clrSet}}

\code{SASObjRef} method: \code{\link[=clrSet,SASObjRef-method]{clrSet}}
}
\keyword{internal}
