% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{cnx<-,SASEGDriver-method}
\alias{cnx<-,SASEGDriver-method}
\title{Set the value of slot cnx of a SASEGDriver object}
\usage{
\S4method{cnx}{SASEGDriver}(obj) <- value
}
\arguments{
\item{obj}{A \code{\linkS4class{SASEGDriver}} object.}

\item{value}{A \code{\linkS4class{SASEGConnection}} or \code{NULL}.}
}
\description{
Value stored in slot \code{cnx} can be replaced with this method. If there 
    is an active connection, the only value accepted is \code{NULL}, an error 
    is raised otherwise. If there is no active connection, an object of class 
    \code{\linkS4class{SASEGConnection}} is accepted as a value. This method 
    is not exported. Only developpers may need to use it.
}
\seealso{
dbListConnections,SASEGDriver-method
}
\keyword{internal}
