% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{cnx<-}
\alias{cnx<-}
\title{Replace the value of slot cnx}
\usage{
cnx(obj) <- value
}
\arguments{
\item{obj}{An object with a slot named \code{cnx}.}

\item{value}{A value to replace for.}
}
\description{
A replacement method for a slot named \code{cnx}.
}
\keyword{internal}
