% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{countOutputDatasets}
\alias{countOutputDatasets}
\title{Get the number of datasets generated by the code}
\usage{
countOutputDatasets(code)
}
\description{
\code{countOutputDatasets} method is used to get the number of datasets 
    created by the code execution.
}
\keyword{internal}
