% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{dataset}
\alias{dataset}
\title{Refer to a SAS dataset}
\usage{
dataset(name, libname = "WORK")
}
\arguments{
\item{name}{A character string with the dataset name. Compulsory.}

\item{libname}{A character string with the libname. Optional.}
}
\value{
\code{dataset} function returns a \code{\link[DBI]{Table-class}} object.
}
\description{
\code{dataset} function is used to refer to a \code{SAS} dataset.
}
\examples{
dataset("SASHELP.CLASS")
dataset(libname = "SASHELP", name = "CLASS")

# For a dataset in WORK library, you can use:
dataset("TEST")
}
