% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbClearListResults,SASEGConnection-method}
\alias{dbClearListResults,SASEGConnection-method}
\title{Clear list of results}
\usage{
\S4method{dbClearListResults}{SASEGConnection}(conn, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{SASEGConnection}} object.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
This method removes invalid \code{\linkS4class{SASEGSQLResult}} of the list 
of results referenced in a \code{\linkS4class{SASEGConnection}} object. This
method is not exported.
}
\keyword{internal}
