% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{dbClearListResults}
\alias{dbClearListResults}
\title{Clear list of results}
\usage{
dbClearListResults(conn, ...)
}
\arguments{
\item{conn}{An object.}

\item{...}{Other parameters passed on to method.}
}
\description{
This method removes invalid results from the list of results.
}
\keyword{internal}
