% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbClearResult,SASEGSQLResult-method}
\alias{dbClearResult,SASEGSQLResult-method}
\title{Clear a result object}
\usage{
\S4method{dbClearResult}{SASEGSQLResult}(res, ...)
}
\arguments{
\item{res}{An object resulting from a call to 
\code{\link[=dbSendQuery,SASEGConnection,character-method]{dbSendQuery}} or
\code{\link[=dbSendStatement,SASEGConnection,character-method]{dbSendStatement}}.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
\code{dbClearResult} method drops temporary datasets created by a call to 
\code{\link[=dbSendQuery,SASEGConnection,character-method]{dbSendQuery}} or
\code{\link[=dbSendStatement,SASEGConnection,character-method]{dbSendStatement}}.
}
\seealso{
Generic: \code{\link[DBI]{dbClearResult}}
}
