% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbColumnInfo,SASEGSQLResult-method}
\alias{dbColumnInfo,SASEGSQLResult-method}
\title{Get information about result types}
\usage{
\S4method{dbColumnInfo}{SASEGSQLResult}(res, ...)
}
\arguments{
\item{res}{An object resulting from a call to 
\code{\link[=dbSendQuery,SASEGConnection,character-method]{dbSendQuery}}.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
A \code{data.frame}.
}
\description{
\code{dbColumnInfo} produces a \code{data.frame} that describes the result 
of a query. Each row of the \code{data.frame} describes a column of the
result dataset and columns in the \code{data.frame} describes the following
aspects of the result dataset:
\itemize{
   \item \code{name}: column name (case as-is from query).
   \item \code{field.type}: \code{SAS} column type (char or num).
   \item \code{field.format}: \code{SAS} column format.
   \item \code{field.informat}: \code{SAS} column informat.
   \item \code{notnull}: not \code{NULL} ?
   \item \code{precision}: precision.
   \item \code{scale}: scale.}
}
