% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbConnect,SASEGConnection-method}
\alias{dbConnect,SASEGConnection-method}
\title{Replicate a connection}
\usage{
\S4method{dbConnect}{SASEGConnection}(drv, ...)
}
\arguments{
\item{drv}{An object created with \code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
A \code{\linkS4class{SASEGConnection}} object.
}
\description{
\code{dbConnect} method replicates a connection. As 
\code{\linkS4class{SASEGDriver}} class allows a single connection per 
driver object, the behavior of \code{dbConnect} is the following: 
\itemize{
\item If the \code{\linkS4class{SASEGConnection}} object passed in argument 
is active, \code{dbConnect} returns this object. 
\item If the \code{\linkS4class{SASEGConnection}} object passed in argument 
is inactive, \code{dbConnect} returns a new valid connection with same 
profile and server.}
}
\seealso{
Generic: \code{\link[DBI]{dbConnect}}.

Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListResults,SASEGConnection-method}},
  \code{\link{dbListTables,SASEGConnection-method}},
  \code{\link{dbRemoveTable,SASEGConnection-method}}
}
