% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbConnectString,SASEGConnection-method}
\alias{dbConnectString,SASEGConnection-method}
\title{Get DBMS connect string for SAS/ACCESS}
\usage{
\S4method{dbConnectString}{SASEGConnection}(conn, cnx.alias = "CNX", ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{SASEGConnection}} object.}

\item{cnx.alias}{A character string with the alias name for the connection.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
A character string.
}
\description{
This method returns a character string with the connection string for 
\code{SAS/ACCESS}. This method is not exported.
}
\keyword{internal}
