% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{dbConnectString}
\alias{dbConnectString}
\title{Get DBMS connect string}
\usage{
dbConnectString(conn, ...)
}
\arguments{
\item{conn}{An object.}

\item{...}{Other parameters passed on to method.}
}
\description{
Get DBMS connect string
}
\keyword{internal}
