% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbDataType,SASEGConnection-method}
\alias{dbDataType,SASEGConnection-method}
\title{Find the SAS data type associated with an R object}
\usage{
\S4method{dbDataType}{SASEGConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{An object of class \code{\linkS4class{SASEGConnection}}.}

\item{obj}{An \code{R} object whose \code{SAS} type we want to determine.}

\item{...}{Other parameters. Not used.}
}
\value{
A character string or a character vector.
}
\description{
Find the \code{SAS} data type associated with an \code{R} object.
}
\seealso{
\code{\link{dbDataType,SASEGDriver-method}}.
}
\keyword{internal}
