% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbDataType,SASEGDriver-method}
\alias{dbDataType,SASEGDriver-method}
\title{Find the database data type associated with an R object}
\usage{
\S4method{dbDataType}{SASEGDriver}(dbObj, obj, ...)
}
\description{
Find the database data type associated with an R object
}
