% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbDisconnect,SASEGConnection-method}
\alias{dbDisconnect,SASEGConnection-method}
\title{Disconnect (or close) a SAS EG connection}
\usage{
\S4method{dbDisconnect}{SASEGConnection}(conn, projectPath = NULL, ...)
}
\arguments{
\item{conn}{An object created by \code{\link[=dbConnect,SASEGDriver-method]{dbConnect}}.}

\item{projectPath}{A character string with the path to save the project 
created by \code{RSASEG}. \strong{Be careful: \code{dbDisconnect} method 
overwrites existing files without confirmation}. If \code{NULL}, no project is saved.}
}
\value{
\code{dbDisconnect} returns \code{TRUE}, invisibly.
}
\description{
\code{dbDisconnect} closes a \code{SAS EG} connection. It is important to 
   close connection because creation of a new \code{SASEGConnection} launchs 
   \code{SAS EG} in memory. You also can save your work in a \code{SAS EG} project.
}
\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)

my_profile <- "PROFILE"
my_server <- "SASPROD"
conn <- dbConnect(drv, my_profile, my_server)

dbWriteTable(conn, "mtcars", mtcars)
dbGetQuery(conn, "SELECT * FROM mtcars WHERE cyl = 4")

# Important: you have to disconnect from SAS EG
# When disconnecting, you also can save your work
RSASEG_project <- paste(normalizePath("~"), "RSASEG.egp", sep = "\\\\")
dbDisconnect(conn, projectPath = RSASEG_project)
}
}
\seealso{
Generic: \code{\link[DBI]{dbDisconnect}}.

Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListResults,SASEGConnection-method}},
  \code{\link{dbListTables,SASEGConnection-method}},
  \code{\link{dbRemoveTable,SASEGConnection-method}}
}
