% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbExistsTable,SASEGConnection-method}
\alias{dbExistsTable,SASEGConnection-method}
\title{Does a dataset exist on a SAS server ?}
\usage{
\S4method{dbExistsTable}{SASEGConnection}(conn, name, ...)
}
\arguments{
\item{conn}{An object returned by \code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{name}{A character string with the name of a dataset (one-level or 
two-levels names). One-level dataset names are interpreted as \code{WORK} 
library datasets.}

\item{...}{Other parameters passed on. Not used.}
}
\value{
A logical.
}
\description{
\code{dbExistsTable} is used to test if a dataset exists on a \code{SAS} server.
}
\seealso{
Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListResults,SASEGConnection-method}},
  \code{\link{dbListTables,SASEGConnection-method}}
}
