% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbFetch,SASEGResult-method}
\alias{dbFetch,SASEGResult-method}
\title{Retrieve records from SAS EG query}
\usage{
\S4method{dbFetch}{SASEGResult}(res, n = -1, ...)
}
\description{
Retrieve records from SAS EG query
}
