% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{dbFetchAll}
\alias{dbFetchAll}
\title{Fetch all result tables}
\usage{
dbFetchAll(res, ...)
}
\arguments{
\item{res}{A result object.}

\item{...}{Other parameters passed on to method.}
}
\description{
In case of a program creates multiple tables, fetch all tables.
}
