% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetException,SASEGConnection-method}
\alias{dbGetException,SASEGConnection-method}
\title{Get SAS/PROC SQL exceptions}
\usage{
\S4method{dbGetException}{SASEGConnection}(conn, ...)
}
\arguments{
\item{conn}{An object created by \code{\link[=dbConnect,SASEGDriver-method]{dbConnect}}.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
A list with elements \code{errorNum} (an integer error number) and 
    \code{errorMsg} (a character string) describing the last error in the 
    connection \code{conn}.
}
\description{
Get the last \code{SAS/PROC SQL} exception.
}
\details{
\code{errorNum} and \code{errorMsg} are given by the \code{SAS} automatic
    macro variables \code{SQLRC} and \code{SYSERRORTEXT}.
}
\seealso{
Generic: \code{\link[DBI]{dbGetException}}.

Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListResults,SASEGConnection-method}},
  \code{\link{dbListTables,SASEGConnection-method}},
  \code{\link{dbRemoveTable,SASEGConnection-method}}
}
