% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetInfo,SASEGConnection-method}
\alias{dbGetInfo,SASEGConnection-method}
\title{Get informations about a connection}
\usage{
\S4method{dbGetInfo}{SASEGConnection}(dbObj, ...)
}
\description{
\code{dbGetInfo} returns informations about a \code{\linkS4class{SASEGConnection}}
object:
\itemize{
\item \code{profile}: profile used in connection.
\item \code{db.version}: version number of \code{SAS} (this is the 
    \code{SYSVER} \code{SAS} macro variable).
\item \code{dbname}: \code{USER} libname, default: \code{WORK} (this is the 
    \code{SAS} \code{USER} option).
\item \code{username}: username on \code{SAS} server (this is the 
    \code{SYSUSERID} \code{SAS} macro variable).
\item \code{host}: host name.
\item \code{port}: port number.}
}
\seealso{
Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection,character-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListResults,SASEGConnection-method}},
  \code{\link{dbListTables,SASEGConnection-method}}
}
