% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetInfo,SASEGDriver-method}
\alias{dbGetInfo,SASEGDriver-method}
\title{Informations about a SASEGDriver}
\usage{
\S4method{dbGetInfo}{SASEGDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object created by a call to \code{\link[=SASEG]{SASEG()}}.}

\item{...}{Other parameters passed on to method. Not used.}
}
\description{
\code{dbGetInfo} method is called to get:
\itemize{
    \item \code{driver.version}: version of \code{RSASEG} package.
    \item \code{client.version}: version of \code{SAS Enterprise Guide}.
    \item \code{max.connections}: equals to 1.
    \item \code{available.profiles}: a list with informations about 
        available profiles.
}
}
\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)

dbGetInfo(drv)
dbUnloadDriver(drv)}
}
\seealso{
\code{SASEGDriver} methods: 
    \code{\link[=dbIsValid,SASEGDriver-method]{dbIsValid}},
    \code{\link[=dbListConnections,SASEGDriver-method]{dbListConnections}},
    \code{\link[=dbUnloadDriver,SASEGDriver-method]{dbUnloadDriver}},
    \code{\link[=dbDataType,SASEGDriver-method]{dbDataType}}
}
