% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetLog,SASEGResult-method}
\alias{dbGetLog,SASEGResult-method}
\title{Get the log ater a SAS run}
\usage{
\S4method{dbGetLog}{SASEGResult}(res, ...)
}
\arguments{
\item{res}{A \code{SASEGResult}.}
}
\description{
\code{dbGetLog} retrieves the log of the \code{SAS} program.
}
