% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{dbGetLog}
\alias{dbGetLog}
\title{Get the log of a program execution}
\usage{
dbGetLog(res, ...)
}
\description{
\code{dbGetLog} methods retrieve the log of a program execution.
}
\details{
\code{dbGetLog} does not belong to the \code{DBI} specification.
}
\keyword{internal}
