% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetQuery,SASEGConnection,character-method}
\alias{dbGetQuery,SASEGConnection,character-method}
\title{Send an SQL query to SAS EG and get results}
\usage{
\S4method{dbGetQuery}{SASEGConnection,character}(conn, statement,
  codeName = NULL, persistent = TRUE, ...)
}
\arguments{
\item{conn}{A \code{SASEGConnection} object.}

\item{statement}{A character string containing an \code{SQL} code 
or an \code{\link[DBI]{SQL}} class object.}

\item{codeName}{A character string to name the new \code{\linkS4class{SASEGCode}} object.}

\item{persistent}{A logical. If \code{TRUE}, a new 
\code{\linkS4class{SASEGCode}} is created in the \code{SASEGProject}. If 
\code{FALSE}, \code{garbage} code is used.}

\item{...}{Other parameters passed on to methods. Not used.}
}
\value{
A \code{data.frame} object.
}
\description{
\code{dbGetQuery} sends an \code{\link[DBI]{SQL}} query to \code{SAS} and
    retrieve results. The query is first embedded in a \code{PROC SQL} and 
    sent to \code{SAS}. \code{dbGetQuery} is used to send \code{SELECT ...} 
    queries.
}
