% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetQuery,SASEGConnection,SAS-method}
\alias{dbGetQuery,SASEGConnection,SAS-method}
\title{Send a SAS program and get result datasets}
\usage{
\S4method{dbGetQuery}{SASEGConnection,SAS}(conn, statement, codeName = NULL,
  persistent = TRUE, ...)
}
\arguments{
\item{conn}{A \code{SASEGConnection} object.}

\item{statement}{A \code{\linkS4class{SAS}} object.}

\item{codeName}{A character string to name the new \code{\linkS4class{SASEGCode}} object.}

\item{persistent}{A logical. If \code{TRUE}, a new 
\code{\linkS4class{SASEGCode}} is created in the \code{SASEGProject}. If 
\code{FALSE}, \code{garbage} code is used.}
}
\value{
\code{dbGetQuery} returns a named (with filenames) list of 
    \code{\link[data.table]{data.table}}.
}
\description{
\code{dbGetQuery} method is used to send a \code{SAS} program and get a list 
of result datasets.
}
