% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetRowsAffected,SASEGSQLResult-method}
\alias{dbGetRowsAffected,SASEGSQLResult-method}
\title{Get the number of rows affected}
\usage{
\S4method{dbGetRowsAffected}{SASEGSQLResult}(res, ...)
}
\arguments{
\item{res}{An object resulting from a call to 
\code{\link[=dbSendStatement,SASEGConnection,character-method]{dbSendStatement}}.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
Number of rows affected.
}
\description{
\code{dbGetRowsAffected} returns the number of rows that were added, deleted,
or updated by a data manipulation statement.
}
\seealso{
Generic: \code{\link[DBI]{dbGetRowsAffected}}.
}
