% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetStatement,SASEGSQLResult-method}
\alias{dbGetStatement,SASEGSQLResult-method}
\title{Get the statement associated with a result object}
\usage{
\S4method{dbGetStatement}{SASEGSQLResult}(res, ...)
}
\arguments{
\item{res}{An object resulting from a call to 
\code{\link[=dbSendQuery,SASEGConnection,character-method]{dbSendQuery}} or
\code{\link[=dbSendStatement,SASEGConnection,character-method]{dbSendStatement}}.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
A character string.
}
\description{
\code{dbGetStatement} returns the statement that was passed on to 
    \code{\link[=dbSendQuery,SASEGConnection,character-method]{dbSendQuery}} or
    \code{\link[=dbSendStatement,SASEGConnection,character-method]{dbSendStatement}}.
}
\seealso{
Generic: \code{\link[DBI]{dbGetStatement}}
}
