% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbHasCompleted,SASEGResult-method}
\alias{dbHasCompleted,SASEGResult-method}
\title{Test if a SAS EG result is fetched}
\usage{
\S4method{dbHasCompleted}{SASEGResult}(res, ...)
}
\arguments{
\item{res}{An object of class \code{\linkS4class{SASEGResult}} or that 
inherits from this class, as \code{\linkS4class{SASEGSQLResult}}.}

\item{...}{Other parameters. Not used.}
}
\value{
A logical: \code{TRUE}, if all rows are fetched; \code{FALSE}, if 
    there is some row to be fetched.
}
\description{
\code{dbHasCompleted} method is used to test if a result is totally fetched.
}
\seealso{
Generic: \code{\link[DBI]{dbHasCompleted}}.
}
\keyword{internal}
