% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbIsValid,SASEGDriver-method}
\alias{dbIsValid,SASEGDriver-method}
\title{Is this SASEGDriver still valid ?}
\usage{
\S4method{dbIsValid}{SASEGDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object created by a call to \code{\link[=SASEG]{SASEG()}}.}

\item{...}{Other parameters passed on to method. Not used.}
}
\value{
\code{dbIsValid} returns a \code{logical}.
}
\description{
\code{dbIsValid} tests if a \code{SASEGDriver} object is still valid. A call
    to \code{\link[=dbUnloadDriver,SASEGDriver-method]{dbUnloadDriver}} 
    invalidates the object.
}
\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)
dbIsValid(drv)

dbUnloadDriver(drv)
dbIsValid(drv)}
}
\seealso{
\code{SASEGDriver} methods: 
    \code{\link[=dbGetInfo,SASEGDriver-method]{dbGetInfo}},
    \code{\link[=dbListConnections,SASEGDriver-method]{dbListConnections}},
    \code{\link[=dbUnloadDriver,SASEGDriver-method]{dbUnloadDriver}},
    \code{\link[=dbDataType,SASEGDriver-method]{dbDataType}}
}
