% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbIsValid,SASEGResult-method}
\alias{dbIsValid,SASEGResult-method}
\title{Test if a SASEGResult object is valid}
\usage{
\S4method{dbIsValid}{SASEGResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{A \code{\linkS4class{SASEGResult}} object.}

\item{...}{Other parameters. Not used.}
}
\value{
A logical.
}
\description{
\code{dbIsValid} method tests if a \code{\linkS4class{SASEGResult}} is valid.
\code{TRUE} is returned if the object is valid.
}
\seealso{
Generic: \code{\link[DBI]{dbIsValid}}.
}
