% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbListConnections,SASEGDriver-method}
\alias{dbListConnections,SASEGDriver-method}
\title{List active SAS EG connections}
\usage{
\S4method{dbListConnections}{SASEGDriver}(drv, ...)
}
\arguments{
\item{drv}{An object created by a call to \code{\link[=SASEG]{SASEG()}}.}

\item{...}{Other parameters. Not used.}
}
\value{
\code{dbListConnections} returns an empty list if there is no active
    connection or a list with a single \code{\linkS4class{SASEGConnection}} 
    object.
}
\description{
\code{dbListConnections} method is used to retrieve the list of active 
    connections. Since \code{SASEGDriver} allows a single connection, the 
    list contains either a single element or is empty.
}
\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)

dbListConnections(drv)

my_profile <- "PROFILE"
my_server <- "SASPROD"
conn <- dbConnect(drv, my_profile, my_server)
dbListConnections(drv)

dbDisconnect(conn)
dbListConnections(drv)

dbUnloadDriver(drv)}
}
\seealso{
\code{SASEGDriver} methods: 
    \code{\link[=dbGetInfo,SASEGDriver-method]{dbGetInfo}},
    \code{\link[=dbIsValid,SASEGDriver-method]{dbIsValid}},
    \code{\link[=dbUnloadDriver,SASEGDriver-method]{dbUnloadDriver}},
    \code{\link[=dbDataType,SASEGDriver-method]{dbDataType}}
}
