% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbListFields,SASEGConnection,character-method}
\alias{dbListFields,SASEGConnection,character-method}
\title{List columns of a dataset}
\usage{
\S4method{dbListFields}{SASEGConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{An object returned by \code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{name}{A character string with the name of a dataset (one-level or 
two-levels names). One-level dataset names are interpreted as \code{WORK} 
library datasets.}

\item{...}{Other parameters passed on. Not used.}
}
\value{
A character vector.
}
\description{
List column names of a remote dataset.
}
\seealso{
Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection,character-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListResults,SASEGConnection-method}},
  \code{\link{dbListTables,SASEGConnection-method}}
}
