% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbListResults,SASEGConnection-method}
\alias{dbListResults,SASEGConnection-method}
\title{List results}
\usage{
\S4method{dbListResults}{SASEGConnection}(conn, ...)
}
\arguments{
\item{conn}{An object returned by \code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{...}{Other parameters passed on. Not used.}
}
\value{
A list of \code{\linkS4class{SASEGSQLResult}} objects.
}
\description{
List valid \code{\linkS4class{SASEGSQLResult}} objects.
}
\seealso{
Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection,character-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListTables,SASEGConnection-method}}
}
