% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbListTables,SASEGConnection-method}
\alias{dbListTables,SASEGConnection-method}
\title{List datasets available on a SAS server}
\usage{
\S4method{dbListTables}{SASEGConnection}(conn, ...)
}
\arguments{
\item{conn}{An object returned by \code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{...}{Other parameters passed on. Not used.}
}
\value{
A character vector.
}
\description{
List datasets available on a \code{SAS} server.
}
\seealso{
Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListResults,SASEGConnection-method}}
}
