% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbQuoteIdentifier,SASEGConnection,character-method}
\alias{dbQuoteIdentifier,SASEGConnection,character-method}
\alias{dbQuoteIdentifier,SASEGConnection,SQL-method}
\alias{dbQuoteIdentifier,SASEGConnection,Table-method}
\title{Quote identifiers}
\usage{
\S4method{dbQuoteIdentifier}{SASEGConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{SASEGConnection,SQL}(conn, x, ...)

\S4method{dbQuoteIdentifier}{SASEGConnection,Table}(conn, x, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{SASEGConnection}} object.}

\item{x}{A character vector to quote as identifier.}

\item{...}{Other arguments passed on to methods.}
}
\value{
An object that can be coerced to character of the same length as the 
    input.
}
\description{
Quote identifiers.
}
\seealso{
Generic: \code{\link[DBI]{dbQuoteIdentifier}}.
}
