% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbQuoteString,SASEGConnection,character-method}
\alias{dbQuoteString,SASEGConnection,character-method}
\alias{dbQuoteString,SASEGConnection,SQL-method}
\title{Quote literal strings}
\usage{
\S4method{dbQuoteString}{SASEGConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{SASEGConnection,SQL}(conn, x, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{SASEGConnection}} object.}

\item{x}{A character vector to quote as string.}

\item{...}{Other arguments passed on to methods.}
}
\value{
An object that can be coerced to character of the same length as the 
    input.
}
\description{
Quote literal strings.
}
\seealso{
Generic: \code{\link[DBI]{dbQuoteString}}.
}
