% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbReadTable,SASEGConnection,character-method}
\alias{dbReadTable,SASEGConnection,character-method}
\title{Copy data frames from SAS dataset}
\usage{
\S4method{dbReadTable}{SASEGConnection,character}(conn, name, ...,
  row.names = NA, check.names = TRUE)
}
\arguments{
\item{conn}{An object returned by 
\code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{name}{A character string specifying the unquoted dataset name, or the 
result of a call to 
\code{\link[=dbQuoteIdentifier,SASEGConnection,character-method]{dbQuoteIdentifier()}}}

\item{...}{Other parameters passed on to methods. Not used.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{check.names}{A logical. Check if column names are valid \code{R} identifiers.}
}
\description{
Reads a \code{SAS} dataset to a data frame, optionally converting a column 
to row names and converting the column names to valid \code{R} identifiers.
}
\seealso{
Generic: \code{\link[DBI]{dbReadTable}}.
}
