% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbRemoveTable,SASEGConnection-method}
\alias{dbRemoveTable,SASEGConnection-method}
\title{Remove a dataset on a SAS server}
\usage{
\S4method{dbRemoveTable}{SASEGConnection}(conn, name, codeName = NULL,
  persistent = TRUE, ...)
}
\arguments{
\item{conn}{An object returned by \code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{name}{A character string with the name of a dataset (one-level or 
two-levels names). One-level dataset names are interpreted as \code{WORK} 
library datasets.}

\item{codeName}{A character string to name the new \code{\linkS4class{SASEGCode}} object.}

\item{persistent}{A logical. If \code{TRUE}, a new 
\code{\linkS4class{SASEGCode}} is created in the \code{SASEGProject}. If 
\code{FALSE}, \code{garbage} code is used.}

\item{...}{Other parameters passed on. Not used.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
\code{dbRemoveTable} method is used to drop a dataset.
}
\seealso{
Generic: \code{\link[DBI]{dbRemoveTable}}.

Other SASEGConnection class methods: \code{\link{SASEG}},
  \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListResults,SASEGConnection-method}},
  \code{\link{dbListTables,SASEGConnection-method}}
}
