% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbSendQuery,SASEGConnection,character-method}
\alias{dbSendQuery,SASEGConnection,character-method}
\title{Send an SQL query to SAS EG}
\usage{
\S4method{dbSendQuery}{SASEGConnection,character}(conn, statement,
  persistent = TRUE, codeName = NULL, query = TRUE, ...)
}
\arguments{
\item{conn}{A \code{SASEGConnection} object.}

\item{statement}{A character string containing a \code{SQL} code or an 
\code{\link[DBI]{SQL}} class object.}

\item{persistent}{A logical. If \code{TRUE}, a new \code{SASEGCode} 
is created in the \code{SASEGProject}. If \code{FALSE}, \code{garbage} 
code is used.}

\item{codeName}{A character string to name the new \code{\linkS4class{SASEGCode}} object.}

\item{query}{A logical. \code{TRUE} indicates a \code{SELECT} query. 
\code{FALSE} indicate a data transformation statement.}
}
\value{
A \code{SASEGSQLResult} object.
}
\description{
\code{dbSendQuery} sends an \code{SQL} query to \code{SAS}. The query is 
    first embedded in a \code{PROC SQL} and sent to \code{SAS}. 
    \code{dbSendQuery} is used to send \code{SELECT ...} queries.
}
\examples{
# This is another good place to put examples
}
\seealso{
Generic: \code{\link[DBI]{dbSendQuery}}
}
