% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbSendQuery,SASEGConnection-method}
\alias{dbSendQuery,SASEGConnection-method}
\title{Send a query to SAS EG.}
\usage{
\S4method{dbSendQuery}{SASEGConnection}(conn, statement, codeName = NULL, ...)
}
\description{
Send a query to SAS EG.
}
\examples{
# This is another good place to put examples
}
