% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbSendStatement,SASEGConnection,character-method}
\alias{dbSendStatement,SASEGConnection,character-method}
\title{Send an SQL statement to SAS EG}
\usage{
\S4method{dbSendStatement}{SASEGConnection,character}(conn, statement,
  codeName = NULL, persistent = TRUE, send.dbms = TRUE, ...)
}
\arguments{
\item{conn}{A \code{SASEGConnection} object.}

\item{statement}{A character string containing an \code{SQL} code 
or an \code{\link[DBI]{SQL}} class object.}

\item{codeName}{A character string to name the new \code{\linkS4class{SASEGCode}} object.}

\item{persistent}{A logical. If \code{TRUE}, a new 
\code{\linkS4class{SASEGCode}} is created in the \code{SASEGProject}. If 
\code{FALSE}, \code{garbage} code is used.}

\item{...}{Other parameters passed on to methods. Not used.}
}
\value{
A \code{\linkS4class{SASEGSQLResult}} object.
}
\description{
\code{dbSendStatement} sends an \code{\link[DBI]{SQL}} data manipulation 
    statement to \code{SAS}. The query is first embedded in a 
    \code{PROC SQL} and sent to \code{SAS}.
}
\seealso{
\code{\link[=dbSendQuery,SASEGConnection,character-method]{dbSendQuery()}}

Generic: \code{\link[DBI]{dbSendStatement}}
}
