% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbUnloadDriver,SASEGDriver-method}
\alias{dbUnloadDriver,SASEGDriver-method}
\title{Unload SASEGDriver}
\usage{
\S4method{dbUnloadDriver}{SASEGDriver}(drv, ...)
}
\arguments{
\item{drv}{An object created by a call to \code{\link[=SASEG]{SASEG()}}.}

\item{...}{Other parameters. Not used.}
}
\value{
\code{dbUnloadDriver} method returns \code{TRUE}.
}
\description{
\code{dbUnloadDriver} usage is recommended for \code{SASEGDriver} objects. 
    Since this method is deprecated, a finalizer is registered upon garbage
    collector and at the end of the \code{R} session. The finalizer sends an 
    informative message to the user.
}
\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)

my_profile <- "PROFILE"
my_server <- "SASPROD"
conn <- dbConnect(drv, my_profile, my_server)

dbWriteTable(conn, "mtcars", mtcars)
dbGetQuery(conn, "SELECT * FROM mtcars WHERE cyl = 4")

# Important: you have to disconnect from SAS EG and
#            unload driver.
# 
# When disconnecting, you also can save your work to 
# a SAS EG Project file.

dbDisconnect(conn)
dbUnloadDriver(drv)}
}
\seealso{
\code{SASEGDriver} methods: 
    \code{\link[=dbGetInfo,SASEGDriver-method]{dbGetInfo}},
    \code{\link[=dbIsValid,SASEGDriver-method]{dbIsValid}},
    \code{\link[=dbListConnections,SASEGDriver-method]{dbListConnections}},
    \code{\link[=dbDataType,SASEGDriver-method]{dbDataType}}
}
