% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbWriteTable,SASEGConnection-method}
\alias{dbWriteTable,SASEGConnection-method}
\title{Copy data frames to SAS datasets}
\usage{
\S4method{dbWriteTable}{SASEGConnection}(conn, name, value, row.names = NA,
  overwrite = FALSE, append = FALSE, field.types = NULL,
  temporary = FALSE, persistent = TRUE, ...)
}
\arguments{
\item{conn}{An object returned by 
\code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{name}{A character string specifying the unquoted dataset name, or the 
result of a call to 
\code{\link[=dbQuoteIdentifier,SASEGConnection,character-method]{dbQuoteIdentifier()}}}

\item{value}{a \link{data.frame} (or coercible to data.frame).}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{overwrite}{A logical. If \code{TRUE}, an existing dataset of the same 
name will be overwritten.}

\item{append}{A logical. If \code{TRUE}, the rows in an existing table are 
preserved, and the new data are appended.}

\item{field.types}{A named character vector with \code{SAS} data types.}

\item{temporary}{A logical. This is part of the \code{DBI} specification. 
However, this functionality is not supported: one-level named datasets
are created in \code{SAS WORK}. Use two-levels names to permanently save
a dataset.}

\item{persistent}{A logical. If \code{TRUE}, a new 
\code{\linkS4class{SASEGCode}} is created in the \code{SASEGProject}. If 
\code{FALSE}, \code{garbage} code is used.}

\item{...}{Other parameters passed on to methods. Not used.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Writes, overwrites or appends a data frame to a \code{SAS} dataset, 
optionally converting row names to a column and specifying \code{SAS} data
types for fields.
}
\seealso{
Generic: \code{\link[DBI]{dbWriteTable}}.
}
