% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{drv,SASEGConnection-method}
\alias{drv,SASEGConnection-method}
\title{Access to slot drv of a SASEGConnection object}
\usage{
\S4method{drv}{SASEGConnection}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{SASEGconnection}}.}

\item{...}{Other parameters passed to method. Not used.}
}
\value{
A \code{\linkS4class{SASEGDriver}} object.
}
\description{
\code{drv()} method access to the slot \code{drv} of a 
    \code{\linkS4class{SASEGConnection}} object.
}
\keyword{internal}
