% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{drv}
\alias{drv}
\title{Access to slot drv of an object}
\usage{
drv(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Other parameters passed to method.}
}
\description{
A generic accessor to a slot named \code{drv}.
}
\keyword{internal}
