% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{fetched<-}
\alias{fetched<-}
\title{Set the slot fetched as...}
\usage{
fetched(res) <- value
}
\arguments{
\item{res}{A result object.}

\item{value}{A value.}
}
\description{
\code{`fetched<-`} sets the slot \code{fetched} as a value.
}
\keyword{internal}
