% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{finalize_app}
\alias{finalize_app}
\title{A finalizer function for SASEGApplication}
\usage{
finalize_app(e)
}
\arguments{
\item{e}{An external pointer.}
}
\value{
\code{NULL}, invisibly
}
\description{
In case of end user would not safely close a 
    \code{\linkS4class{SASEGApplication}} object, a finalizer is 
    implemented. A call to method \code{"Quit"} is done.
}
\keyword{internal}
