% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{finalize_cnx}
\alias{finalize_cnx}
\title{A finalizer function for SASEGConnection}
\usage{
finalize_cnx(e)
}
\arguments{
\item{e}{An environment. This is the \code{infos} slot of a 
\code{\linkS4class{SASEGConnection}} object.}
}
\value{
\code{NULL}, invisibly.
}
\description{
A finaliser function for \code{\linkS4class{SASEGConnection}} objects. 
This function is not exported.
}
\keyword{internal}
