% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{getDBMSArgs}
\alias{getDBMSArgs}
\title{Get DBMS arguments slot of an object}
\usage{
getDBMSArgs(obj)
}
\arguments{
\item{obj}{An object.}
}
\description{
Methods to get DBMS arguments slot of an object.
}
\keyword{internal}
