% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{getFileName}
\alias{getFileName}
\title{Get the filename of an object}
\usage{
getFileName(object, ...)
}
\arguments{
\item{...}{Other parameters passed on to method.}
}
\description{
\code{getFileName} method is used to get the filename of an object.
}
\keyword{internal}
