% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{getListAvailableProfiles}
\alias{getListAvailableProfiles}
\title{Get a list of available profiles}
\usage{
getListAvailableProfiles(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Other parameters passed to method.}
}
\description{
Get a list of available profiles.
}
\keyword{internal}
