% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{getListDatasets}
\alias{getListDatasets}
\title{Get the list of the datasets}
\usage{
getListDatasets(code)
}
\description{
\code{getListsDatasets} method is used to retrieve a list of the dataset(s) 
    created by the code execution.
}
\keyword{internal}
