% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{getLog}
\alias{getLog}
\title{Get the log}
\usage{
getLog(code)
}
\arguments{
\item{code}{A code object.}
}
\description{
\code{getLog} method is used to retrieve the log after the code execution.
}
\keyword{internal}
