% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{getName}
\alias{getName}
\title{Get the name of an object}
\usage{
getName(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Other parameters passed on to method.}
}
\description{
Get the name of an object.
}
\keyword{internal}
