% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{getProfile}
\alias{getProfile}
\title{Get the profile of an object}
\usage{
getProfile(obj)
}
\arguments{
\item{obj}{An object.}
}
\description{
Methods to get profile of an object.
}
\keyword{internal}
