% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{getSourceCode}
\alias{getSourceCode}
\title{Get the source code}
\usage{
getSourceCode(code)
}
\arguments{
\item{code}{A code object.}
}
\description{
\code{getSourceCode} method is used to retrieve the source code after execution.
}
\keyword{internal}
