% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{getVersion}
\alias{getVersion}
\title{Get the version of an object}
\usage{
getVersion(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Other parameters passed to method.}
}
\description{
Get the version of an object.
}
\keyword{internal}
