% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{isValid<-,SASEGConnection-method}
\alias{isValid<-,SASEGConnection-method}
\title{Set the value isValid of a SASEGConnection object}
\usage{
\S4method{isValid}{SASEGConnection}(obj) <- value
}
\arguments{
\item{obj}{A \code{\linkS4class{SASEGDriver}} object.}

\item{value}{A logical.}
}
\description{
Value stored in slot \code{isValid} can be replaced with this method. This 
    method is not exported. Only developpers may need to use it.
}
\seealso{
dbIsValid,SASEGConnection-method
}
\keyword{internal}
