% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{isValid<-}
\alias{isValid<-}
\title{Set the value isValid of an object}
\usage{
isValid(obj) <- value
}
\arguments{
\item{obj}{An object with a slot named \code{isValid}.}

\item{value}{A value to replace for.}
}
\description{
A replacement method for a slot named \code{isValid}.
}
\keyword{internal}
