% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{loadSASEGScripting}
\alias{loadSASEGScripting}
\title{Load SAS EG Scripting API}
\usage{
loadSASEGScripting(DLLFilePath)
}
\arguments{
\item{DLLFilePath}{A string with the path to the \code{SASEGScripting.dll} file.}
}
\description{
Load \code{SAS Enterprise Guide Scripting} DLL.
You have to provide the path to the file \code{SASEGScripting.dll} that comes 
with your installation.
}
\details{
This function is a simple wrapper over \code{\link[rClr]{clrLoadAssembly}}.
}
\examples{
\dontrun{

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
loadSASEGScripting(path)
}
}
\keyword{internal}
