% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{newCode}
\alias{newCode}
\title{Create a new code}
\usage{
newCode(project, ...)
}
\arguments{
\item{project}{A project object.}
}
\description{
\code{newCode} method creates a new code object in an existing project.
}
\keyword{internal}
