% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{newProject}
\alias{newProject}
\title{Create a new project in an application}
\usage{
newProject(application)
}
\arguments{
\item{application}{An application object.}
}
\description{
\code{newProject} method creates a new project in an application.
}
\keyword{internal}
