% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runr.R
\name{proc_saseg}
\alias{proc_saseg}
\title{Runr proc}
\usage{
proc_saseg(DLLPath, profile, server, projectpath = NULL, ...)
}
\arguments{
\item{DLLPath}{A character string with the path to file \code{SASEGScripting.dll}.}

\item{profile}{A character string with the \code{SAS EG} profile name.}

\item{server}{A character string with the \code{SAS} server name to run programs.}

\item{...}{Other parameters. Not used.}
}
\description{
Function to use \code{SAS} engine with runr.
}
\keyword{internal}
