% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{ptr}
\alias{ptr}
\title{Get the ptr slot of an object}
\usage{
ptr(SASObj)
}
\arguments{
\item{SASObj}{An object with a \code{ptr} slot.}
}
\description{
A generic. Get the \code{ptr} slot of an object.
}
\keyword{internal}
