% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{SAS,character-method}
\alias{SAS,character-method}
\title{Quote a character string as a SAS program}
\usage{
\S4method{SAS}{character}(x, ...)
}
\arguments{
\item{x}{A character string.}

\item{...}{Other parameters passed on to methods.}
}
\value{
An object of class \code{SAS}.
}
\description{
This method quotes a character string as a \code{SAS} program.
This method is very similar to \code{\link[DBI]{SQL}} method for class \code{character}.
}
\seealso{
Other SAS-methods: \code{\link{SAS,SQL-method}},
  \code{\link{SAS-class}}
}
\keyword{internal}
