% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{SAS,SQL-method}
\alias{SAS,SQL-method}
\title{Transform an SQL statement as a SAS statement}
\usage{
\S4method{SAS}{SQL}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[DBI]{SQL}}.}
}
\value{
An object of class \code{SAS}.
}
\description{
This method wraps an \code{SQL} statement into a \code{PROC SQL}.
}
\details{
The \code{ODS} option permits to create a \code{SAS} dataset when a 
    \code{SELECT} statement is submitted (thx @ François Malet).
}
\examples{
sql_statement <- DBI::SQL("SELECT * \\n FROM SASHELP.CLASS")
sas_pgm <- RSASEG::SAS(sql_statement)
show(sas_pgm)
}
\seealso{
Other SAS-methods: \code{\link{SAS,character-method}},
  \code{\link{SAS-class}}
}
\keyword{internal}
