% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGConnection-class}
\alias{SASEGConnection-class}
\title{SAS EG connection class}
\description{
This class inherits from \code{\link[DBI]{DBIConnection-class}}.
An object of class \code{SASEGConnection} can be understood a \code{SAS EG} project.
}
\section{Slots}{

\describe{
\item{\code{infos}}{An environment that reference all slots. This environment is 
registered. Referenced objects are: drv (the SASEGDriver used to connect to), 
isValid (a logical), profile (a character string), server (a character 
string), SASProject (a SASEGProject object), SASUtil (a SASEGCode object), 
listResults (an environment), dbms (a character string)}
}}

\keyword{internal}
