% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGResult-class}
\alias{SASEGResult-class}
\alias{fetched<-,SASEGResult-method}
\alias{isValid<-,SASEGResult-method}
\alias{dbIsValid,SASEGResult-method}
\alias{dbHasCompleted,SASEGResult-method}
\alias{dbFetchAll,SASEGResult-method}
\alias{dbGetLog,SASEGResult-method}
\title{SASEGResult class}
\usage{
\S4method{fetched}{SASEGResult}(res) <- value

\S4method{isValid}{SASEGResult}(obj) <- value

\S4method{dbIsValid}{SASEGResult}(dbObj, ...)

\S4method{dbHasCompleted}{SASEGResult}(res, ...)

\S4method{dbFetchAll}{SASEGResult}(res, ...)

\S4method{dbGetLog}{SASEGResult}(res, ...)
}
\arguments{
\item{res}{An object of class \code{SASEGResult} or inheriting from this 
class, as \code{\linkS4class{SASEGSQLResult}}.}

\item{value}{A logical.}

\item{obj}{An object of class \code{SASEGResult} or inheriting from this 
class, as \code{\linkS4class{SASEGSQLResult}}.}

\item{dbObj}{An object of class \code{SASEGResult} or inheriting from this 
class, as \code{\linkS4class{SASEGSQLResult}}.}

\item{...}{Other parameters. Not used.}
}
\value{
\code{dbIsValid} returns a logical.

\code{dbHasCompleted} returns a logical: \code{TRUE}, if all rows are 
    fetched; \code{FALSE}, if there is some rows to be fetched.

\code{dbFetchAll} returns a named (with filenames) list of 
    \code{\link[data.table]{data.table}}.

\code{dbGetLog} returns a character string with the \code{SAS log}.
}
\description{
\code{SASEGResult} class inherits from \code{\link[DBI]{DBIResult-class}}. 
   This class represents results of \emph{any} \code{SAS} program. A 
   \code{SASEGResult} object results from a call to 
   \code{\link[=dbSendQuery,SASEGConnection,SAS-method]{dbSendQuery}} with a 
   \code{\linkS4class{SAS}} statement.  It is extended by 
   \code{\linkS4class{SASEGSQLResult}} for \code{\link[DBI]{SQL}} statements.

\code{dbIsValid} method tests if a \code{SASEGResult} is valid.
\code{TRUE} is returned if the object is valid.

\code{dbHasCompleted} method is used to test if a result is totally fetched.

\code{dbFetchAll} method retrieves all datasets (with all rows) created by a 
    \code{SAS} statement. It returns a named list of 
    \code{\link[data.table]{data.table}}. If there is a single fetched 
    dataset, a \code{\link[data.table]{data.table}} is returned.

\code{dbGetLog} method retrieves the log of the \code{\linkS4class{SAS}} 
program. \code{dbGetLog} does not belong to the \code{DBI} specification.
}
\section{Slots}{

\describe{
\item{\code{conn}}{An object of class \code{\linkS4class{SASEGConnection}}.}

\item{\code{SASResult}}{An object of class \code{\linkS4class{SASEGCode}}.}

\item{\code{fetched}}{A closure.}

\item{\code{isValid}}{A closure.}
}}

\section{Replacement methods}{

\code{`fetched<-`} sets the slot \code{fetched} of a \code{SASEGResult} 
    object as \code{TRUE} or \code{FALSE}. This method is not exported. Only 
    developpers may need to use it.


\code{`isValid<-`} method sets the value \code{isValid}. This method is not 
    exported. Only developpers may need to use it.
}

\seealso{
\code{\linkS4class{SASEGSQLResult}}, 
    \code{\link[=dbSendQuery,SASEGConnection,SAS-method]{dbSendQuery}}

Generic: \code{\link[DBI]{dbIsValid}}.

Generic: \code{\link[DBI]{dbHasCompleted}}.
}
\keyword{internal}
