% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGSQLResult-class}
\alias{SASEGSQLResult-class}
\title{SASEG PROC SQL results class}
\description{
\code{SASEGSQLResult} class inherits from \code{\link[DBI]{DBIResult-class}} and extends \code{\linkS4class{SASEGResult}}. 
   This class represents results of \code{PROC SQL} programs.
}
\section{Slots}{

\describe{
\item{\code{SQLResult}}{A character string. This slot contains the filename of the 
dataset created by the \code{ODS} during a \code{PROC SQL}.}

\item{\code{rowsFetched}}{A closure.}
}}

\keyword{internal}
