% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\docType{class}
\name{SASObjRef-class}
\alias{SASObjRef-class}
\alias{SASObjRef}
\alias{ptr,SASObjRef-method}
\alias{clrCall,SASObjRef-method}
\alias{clrGet,SASObjRef-method}
\alias{clrSet,SASObjRef-method}
\title{An S4 virtual class to represent any SAS EG Scripting object}
\usage{
SASObjRef(obj)

\S4method{ptr}{SASObjRef}(SASObj)

\S4method{clrCall}{SASObjRef}(obj, methodName, ...)

\S4method{clrGet}{SASObjRef}(objOrType, name)

\S4method{clrSet}{SASObjRef}(objOrType, name, value)
}
\arguments{
\item{obj}{A \code{CLR} object. Any other \code{R} object can be passed on argument.}

\item{SASObj}{A \code{SASObjRef} object.}
}
\value{
\code{SASObjRef(obj)} returns a \code{SASObjRef} object if \code{obj} is a \code{CLR} object; returns \code{obj} otherwise.

\code{ptr()} returns a \code{cobjRef} object.
}
\description{
The \code{SASObjRef} S4 class represents any \code{SAS EG Scripting} object.
This class and its methods are devoted to developers.
}
\details{
A \code{SASObjRef} object has one slot, named \code{ptr}, where \code{ptr} is
a \code{cobjRef} object. For more information about \code{CLR} objects and methods, see the \code{rClr} package
documentation.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{A \code{cobjRef} object.}
}}

\examples{
\dontrun{

# Modify the path below following your install:
path <- "C:/Program Files/SAS94/SASEnterpriseGuide/7.1/SASEGScripting.dll"

loadSASEGScripting(path)

app_clrobj <- rClr::clrNew("SAS.EG.Scripting.Application")
app_sasobj <- SASObjRef(app_clrobj)

show(app_clrobj)
show(ptr(app_sasobj))

clrGet(app_sasobj, "Name")

clrCall(app_sasobj, "Quit")
}
}
\keyword{internal}
