% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{saveAs}
\alias{saveAs}
\title{Save an object as...}
\usage{
saveAs(object, ...)
}
\arguments{
\item{object}{An object to save.}

\item{...}{Other parameters passed on to method, including filepath.}
}
\description{
A generic that save an object as...
}
\keyword{internal}
