% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{setFetched,SASEGResult-method}
\alias{setFetched,SASEGResult-method}
\title{Set the slot fetched as...}
\usage{
\S4method{setFetched}{SASEGResult}(res, value)
}
\arguments{
\item{res}{A \code{SASEGResult} object.}

\item{value}{A logical.}
}
\value{
\code{TRUE}, invisible.
}
\description{
This method sets the slot \code{fetched} as \code{TRUE} or \code{FALSE}.
}
\keyword{internal}
