% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{setFetched}
\alias{setFetched}
\title{Set the slot Fetched as...}
\usage{
setFetched(res, value)
}
\arguments{
\item{res}{A result object.}

\item{value}{A value.}
}
\description{
This method sets the slot Fetched as a value.
}
\keyword{internal}
