% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{setName}
\alias{setName}
\title{Set the name of an object}
\usage{
setName(object, name, ...)
}
\arguments{
\item{object}{An object to be named.}

\item{name}{A name.}
}
\description{
\code{setName} method is used to set the name of the object.
}
\keyword{internal}
