% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{setProfile}
\alias{setProfile}
\title{Set the profile in an application}
\usage{
setProfile(application, profile)
}
\arguments{
\item{application}{An application object.}

\item{profile}{A profile.}
}
\description{
\code{setProfile} method is used to set the profile in an application.
}
\keyword{internal}
