% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{setServer}
\alias{setServer}
\title{Set the server}
\usage{
setServer(object, server, ...)
}
\arguments{
\item{object}{An object with a server parameter.}

\item{server}{A server name.}
}
\description{
\code{setServer} method is used to set the server.
}
\keyword{internal}
