% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{setText}
\alias{setText}
\title{Set the text}
\usage{
setText(object, text, ...)
}
\arguments{
\item{object}{An object.}

\item{text}{A text.}
}
\description{
\code{setText} method is used to set the text.
}
\keyword{internal}
