% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{sqlAppendTable,SASEGConnection-method}
\alias{sqlAppendTable,SASEGConnection-method}
\title{Insert rows into a dataset}
\usage{
\S4method{sqlAppendTable}{SASEGConnection}(con, table, values, row.names = NA,
  ...)
}
\arguments{
\item{con}{A \code{\linkS4class{SASEGConnection}} object.}

\item{table}{Name of the dataset.}

\item{values}{A data frame. Factors will be converted to character vectors.
Character vectors will be escaped with \code{\link[=dbQuoteString]{dbQuoteString()}}.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{...}{Other arguments used by individual methods.}
}
\value{
An \code{\link[DBI]{SQL}} object with \code{SQL} code.
}
\description{
\code{sqlAppendTable} generates a single \code{SQL} string that inserts a data 
    frame into an existing \code{SAS} dataset.
}
