% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{sqlData,SASEGConnection-method}
\alias{sqlData,SASEGConnection-method}
\title{Convert a data frame into form suitable for upload to SAS}
\usage{
\S4method{sqlData}{SASEGConnection}(con, value, row.names = NA, ...)
}
\arguments{
\item{con}{A \code{\linkS4class{SASEGConnection}} object.}

\item{value}{A data frame}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{...}{Other arguments used by individual methods.}
}
\value{
A data frame with suitable values for \code{SAS}.
}
\description{
Convert a data frame into form suitable for upload to \code{SAS}.
}
\details{
\code{sqlData}:\itemize{
\item Converts factors to characters
\item Quote all strings
\item Converts logical as integers
\item Convert dates as \code{SAS DATE9} format
\item Convert datetimes as \code{SAS DATETIME} format
\item Convert difftimes as \code{SAS TIME} format
\item Replaces \code{NA} with \code{NULL}
}
}
\seealso{
Generic: \code{\link[DBI]{sqlData}}.
}
\keyword{internal}
