% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{terminate}
\alias{terminate}
\title{Terminate (or quit) an object}
\usage{
terminate(object)
}
\arguments{
\item{object}{An object to close.}
}
\description{
\code{terminate} method is used to close (or quit) an object.
}
\keyword{internal}
