% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\name{app}
\alias{app}
\title{Access to slot app of an object}
\usage{
app(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{Other parameters passed to method.}
}
\description{
A generic accessor to a slot named app.
}
\keyword{internal}
