% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dataset}
\alias{dataset}
\alias{dataset,character-method}
\alias{dataset,Table-method}
\alias{dataset,SQL-method}
\title{Refer to a SAS dataset}
\usage{
dataset(name, ...)

\S4method{dataset}{character}(name, libname = "WORK")

\S4method{dataset}{Table}(name, ...)

\S4method{dataset}{SQL}(name, ...)
}
\arguments{
\item{name}{A dataset name. Compulsory.}

\item{...}{Other argument passed on to method.}

\item{libname}{A character string with the libname. Optional.}
}
\value{
\code{dataset} function returns a \code{\link[DBI]{Table-class}} 
    object when \code{param} is a character string.
}
\description{
\code{dataset} function is used to refer to a \code{SAS} dataset. 
\code{\link[DBI]{SQL-class}} objects and \code{\link[DBI]{Table-class}} 
objects are escaped.
}
\examples{
dataset("SASHELP.CLASS")
dataset(libname = "SASHELP", name = "CLASS")

# For a dataset in WORK library, you can use:
dataset("TEST")
}
