% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbDataType,SASEGDriver-method}
\alias{dbDataType,SASEGDriver-method}
\title{Find the SAS data type associated with an R object}
\usage{
\S4method{dbDataType}{SASEGDriver}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{An object resulting from call to \code{SASEG()}.}

\item{obj}{An \code{R} object whose \code{SAS} type we want to determine.}

\item{...}{Other parameters. Not used.}
}
\value{
A character string or a character vector.
}
\description{
Find the \code{SAS} data type associated with an \code{R} object.
}
\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)

dbDataType(drv, 1:5)
dbDataType(drv, 1)
dbDataType(drv, TRUE)
dbDataType(drv, Sys.Date())
dbDataType(drv, Sys.time())
dbDataType(drv, Sys.time() - as.POSIXct(Sys.Date()))
dbDataType(drv, c("x", "abc"))
dbDataType(drv, list(raw(10), raw(20)))
dbDataType(drv, I(3))
dbDataType(drv, iris)

dbUnloadDriver(drv)}
}
\seealso{
Generic: \code{\link[DBI]{dbDataType}}.
}
\keyword{internal}
