% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbDataType,SASEGDriver-method}
\alias{dbDataType,SASEGDriver-method}
\title{Find the SAS data type associated with an R object}
\usage{
\S4method{dbDataType}{SASEGDriver}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{An object resulting from call to \code{SASEG()}.}

\item{obj}{An \code{R} object whose \code{SAS} type we want to determine.}
}
\description{
Find the SAS data type associated with an R object.
}
\examples{
dbDataType(SASEG(), 1:5)
dbDataType(SASEG(), 1)
dbDataType(SASEG(), TRUE)
dbDataType(SASEG(), Sys.Date())
dbDataType(SASEG(), Sys.time())
dbDataType(SASEG(), Sys.time() - as.POSIXct(Sys.Date()))
dbDataType(SASEG(), c("x", "abc"))
dbDataType(SASEG(), list(raw(10), raw(20)))
dbDataType(SASEG(), I(3))

dbDataType(SASEG(), iris)
}
\seealso{
\code{\link[DBI]{dbDataType}}.
}
