% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbDataType,SASEGDriver-method}
\alias{dbDataType,SASEGDriver-method}
\alias{dbDataType,SASEGConnection-method}
\title{Find the SAS data type associated with an R object}
\usage{
\S4method{dbDataType}{SASEGDriver}(dbObj, obj, ...)

\S4method{dbDataType}{SASEGConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{An object resulting from a call to \code{\link[=SASEG]{SASEG()}} 
or \code{\link[=dbConnect,SASEGDriver-method]{dbConnect()}}.}

\item{obj}{An \code{R} object whose \code{SAS} type we want to determine.}

\item{...}{Other parameters. Not used.}
}
\value{
A character string or a character vector.
}
\description{
\code{dbDataType} method finds the \code{SAS} data type associated with an 
\code{R} object.

Methods are implemented for \code{\linkS4class{SASEGDriver}} and 
    \code{\linkS4class{SASEGConnection}} classes.
}
\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)

dbDataType(drv, 1:5)
dbDataType(drv, 1)
dbDataType(drv, TRUE)
dbDataType(drv, Sys.Date())
dbDataType(drv, Sys.time())
dbDataType(drv, Sys.time() - as.POSIXct(Sys.Date()))
dbDataType(drv, c("x", "abc"))
dbDataType(drv, list(raw(10), raw(20)))
dbDataType(drv, I(3))
dbDataType(drv, iris)

my_profile <- "PROFILE"
my_server <- "SASPROD"
conn <- dbConnect(drv, my_profile, my_server)

dbDataType(conn, 1:5)
dbDataType(conn, 1)
dbDataType(conn, TRUE)
dbDataType(conn, Sys.Date())
dbDataType(conn, Sys.time())
dbDataType(conn, Sys.time() - as.POSIXct(Sys.Date()))
dbDataType(conn, c("x", "abc"))
dbDataType(conn, list(raw(10), raw(20)))
dbDataType(conn, I(3))
dbDataType(conn, iris)

dbDisconnect(conn, projectPath = RSASEG_project)
dbUnloadDriver(drv)}
}
\seealso{
Generic: \code{\link[DBI]{dbDataType}}.

\code{SASEGDriver} methods: 
    \code{\link[=dbGetInfo,SASEGDriver-method]{dbGetInfo}},
    \code{\link[=dbIsValid,SASEGDriver-method]{dbIsValid}},
    \code{\link[=dbListConnections,SASEGDriver-method]{dbListConnections}},
    \code{\link[=dbUnloadDriver,SASEGDriver-method]{dbUnloadDriver}}
}
