% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbDisconnect,SASEGConnection-method}
\alias{dbDisconnect,SASEGConnection-method}
\title{Disconnect (or close) a SAS EG connection}
\usage{
\S4method{dbDisconnect}{SASEGConnection}(conn, projectPath = NULL, ...)
}
\arguments{
\item{conn}{An object created by \code{\link[=dbConnect,SASEGDriver-method]{dbConnect}}.}

\item{projectPath}{A character string with the path to save the project 
created by \code{RSASEG}. \strong{Be careful: \code{saveAs} method 
overwrites existing files without confirmation.}}
}
\value{
\code{dbDisconnect} returns \code{TRUE}.
}
\description{
\code{dbDisconnect} closes a \code{SAS EG} connection. It is important to 
   close connection because creation of a new \code{SASEGConnection} launchs 
   \code{SAS EG} in memory. You also can save your work in a \code{SAS EG} project.
}
