% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbFetch,SASEGSQLResult-method}
\alias{dbFetch,SASEGSQLResult-method}
\title{Retrieve records from SAS EG SQL query}
\usage{
\S4method{dbFetch}{SASEGSQLResult}(res, n = -1, ...)
}
\description{
Retrieve records from SAS EG SQL query
}
