% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetRowCount,SASEGSQLResult-method}
\alias{dbGetRowCount,SASEGSQLResult-method}
\title{Get the number of fetched rows}
\usage{
\S4method{dbGetRowCount}{SASEGSQLResult}(res, ...)
}
\description{
\code{dbGetRowCount} returns the number of rows that was fetched.
}
