% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbGetStatement,SASEGResult-method}
\alias{dbGetStatement,SASEGResult-method}
\title{Get the statement that was sent to SAS}
\usage{
\S4method{dbGetStatement}{SASEGResult}(res, ...)
}
\description{
\code{dbGetStatement} collect the program that was sent to \code{SAS}.
}
