% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbIsValid,SASEGConnection-method}
\alias{dbIsValid,SASEGConnection-method}
\title{Test if SAS EG connection is valid}
\usage{
\S4method{dbIsValid}{SASEGConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object of class \code{\linkS4class{SASEGConnection}}.}

\item{...}{Other parameters. Not used.}
}
\description{
\code{dbIsValid} tests if a \code{\linkS4class{SASEGConnection}} object is valid.
}
