% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbSendQuery,SASEGConnection,SAS-method}
\alias{dbSendQuery,SASEGConnection,SAS-method}
\title{Send a SAS query to SAS EG}
\usage{
\S4method{dbSendQuery}{SASEGConnection,SAS}(conn, statement,
  persistent = TRUE, codeName = NULL, ...)
}
\arguments{
\item{conn}{A \code{SASEGConnection} object.}

\item{statement}{A \code{\linkS4class{SAS}} object.}

\item{persistent}{A logical. If \code{TRUE}, a new \code{SASEGCode} object 
is created. If \code{FALSE}, \code{garbage} code is used.}

\item{codeName}{A character string to name the new \code{\linkS4class{SASEGCode}} object.}
}
\value{
A \code{SASEGResult} object.
}
\description{
This method sends a \code{SAS} query to \code{SAS EG}.
}
\details{
Statements with class \code{\linkS4class{SAS}} are directly sent to the 
    \code{SAS} server through \code{SAS EG} escaping any code transformations.
}
\keyword{internal}
