% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{dbUnloadDriver,SASEGDriver-method}
\alias{dbUnloadDriver,SASEGDriver-method}
\title{Unload SASEGDriver}
\usage{
\S4method{dbUnloadDriver}{SASEGDriver}(drv, ...)
}
\value{
\code{dbUnloadDriver} method returns \code{TRUE}.
}
\description{
This method was developed for \code{DBI} compliance.
}
\keyword{internal}
