% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{SAS,SQL-method}
\alias{SAS,SQL-method}
\title{Transform an SQL statement into a SAS statement}
\usage{
\S4method{SAS}{SQL}(x, SQLResult = "WORK.SQLOUT")
}
\arguments{
\item{x}{An object of class \code{\link[DBI]{SQL}}.}

\item{SQLResult}{A character string with the name of a table to store 
the \code{SAS/ODS} output. If \code{SQLResult} is \code{NA_character_} 
or \code{""}, the \code{ODS} option is not used.}
}
\value{
An object of class \code{SAS}.
}
\description{
This method wraps an \code{SQL} statement in a \code{PROC SQL}. For instance,
\code{SAS(SQL("SELECT * FROM SASHELP.CLASS"))} returns:
\tabular{l}{
   \code{PROC SQL DQUOTE=ANSI;} \cr
   \code{ods output SQL_Results=WORK.SQLOUT;} \cr
   \code{SELECT * FROM SASHELP.CLASS;} \cr
   \code{QUIT;} \cr
}
}
\details{
The \code{ODS} option permits to create a \code{SAS} dataset when a 
    \code{SELECT} query is submitted (thx @ François Malet).
}
\examples{
sql_statement <- DBI::SQL("SELECT * \\n FROM SASHELP.CLASS")
sas_pgm <- RSASEG::SAS(sql_statement)
show(sas_pgm)
}
\seealso{
Other SAS-methods: \code{\link{SAS,character-method}},
  \code{\link{SAS-class}}
}
\keyword{internal}
