% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{methods}
\name{SASEG}
\alias{SASEG}
\alias{dbConnect,SASEGDriver-method}
\title{A driver for SAS Enterprise Guide}
\usage{

  SASEG(DLLPath = "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll")

\S4method{dbConnect}{SASEGDriver}(drv, profile, server, ...)
}
\arguments{
\item{DLLPath}{A character string with the path to file \code{SASEGScripting.dll}.}

\item{drv}{An object created by a call to \code{\link[=SASEG]{SASEG()}}.}

\item{profile}{A character string with the \code{SAS EG} profile name.}

\item{server}{A character string with the \code{SAS} server name to run programs.}

\item{...}{Other parameters. Not used.}
}
\value{
\code{SASEG()} returns a \code{\linkS4class{SASEGDriver}} object.

\code{dbConnect} returns an object of class \code{\linkS4class{SASEGConnection}}.
}
\description{
\code{SASEG()} generates a new \code{SASEG} driver. You 
    have to provide location of file \code{SASEGScripting.dll} following 
    your installation of \code{SAS Enterprise Guide}. A \code{SASEGDriver} 
    object can be understood as a \code{SAS EG} application.

Use \code{dbConnect} to create a new connection to a \code{SAS} 
    server through \code{SAS Enterprise Guide}. A single connection is 
    allowed by a \code{SASEGDriver} object. If needed, use multiple driver 
    objects instead of multiple connections.
}
\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)

my_profile <- "PROFILE"
my_server <- "SASPROD"
conn <- dbConnect(drv, my_profile, my_server)

dbWriteTable(conn, "mtcars", mtcars)
dbGetQuery(conn, "SELECT * FROM mtcars WHERE cyl = 4")

# Important: you have to disconnect from SAS EG and
#            unload driver.
# 
# When disconnecting, you also can save your work to 
# a SAS EG Project file.

RSASEG_project <- paste(normalizePath("~"), "RSASEG.egp", sep = "\\\\")
dbDisconnect(conn, projectPath = RSASEG_project)
dbUnloadDriver(drv)}
}
\seealso{
\code{SASEGDriver} methods: 
    \code{\link[=dbGetInfo,SASEGDriver-method]{dbGetInfo}},
    \code{\link[=dbIsValid,SASEGDriver-method]{dbIsValid}},
    \code{\link[=dbListConnections,SASEGDriver-method]{dbListConnections}},
    \code{\link[=dbUnloadDriver,SASEGDriver-method]{dbUnloadDriver}},
    \code{\link[=dbDataType,SASEGDriver-method]{dbDataType}}

Other SASEGConnection class methods: \code{\link{dbConnect,SASEGConnection-method}},
  \code{\link{dbDisconnect,SASEGConnection-method}},
  \code{\link{dbExistsTable,SASEGConnection-method}},
  \code{\link{dbGetException,SASEGConnection-method}},
  \code{\link{dbGetInfo,SASEGConnection-method}},
  \code{\link{dbIsValid,SASEGConnection-method}},
  \code{\link{dbListFields,SASEGConnection,character-method}},
  \code{\link{dbListResults,SASEGConnection-method}},
  \code{\link{dbListTables,SASEGConnection-method}},
  \code{\link{dbRemoveTable,SASEGConnection-method}}
}
