% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGDriver-class}
\alias{SASEGDriver-class}
\alias{SASEG}
\alias{dbConnect,SASEGDriver-method}
\title{Driver class for SAS Enterprise Guide}
\usage{
SASEG()

\S4method{dbConnect}{SASEGDriver}(drv, DLLFilePath, profile, server, ...)
}
\arguments{
\item{drv}{An object created by \code{SASEG()}.}

\item{DLLFilePath}{A character string with the filepath to \code{SASEGScripting.dll}.}

\item{profile}{A character string with the \code{SAS EG} profile name.}

\item{server}{A character string with the server name to run \code{SAS} programs.}
}
\value{
\code{SASEG()} returns a \code{SASEGDriver} object.

\code{dbConnect} returns an object of class \code{\linkS4class{SASEGConnection}}.
}
\description{
This a driver class for \code{SAS Enterprise Guide}.

\code{SASEG()} generates a new \code{SASEG} driver.

Use \code{dbConnect} to create a new connection to a \code{SAS} 
    server through \code{SAS Enterprise Guide}.
}
\details{
The \code{SASEGDriver} class inherits from the \code{\link[DBI]{DBIDriver-class}}.
}
\examples{
\dontrun{
# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
my_profile <- "PROFILE"
my_server <- "SASPROD"
conn <- dbConnect(RSASEG::SASEG(), 
                  DLLFilePath = path, 
                  profile = my_profile, 
                  server = my_server)
show(conn)
dbWriteTable(conn, "mtcars", mtcars)
dbGetQuery(conn, "SELECT * FROM mtcars WHERE cyl = 4")

# Important: you have to disconnect from SAS EG
# When disconnecting, you also can save your work
RSASEG_project <- paste(normalizePath("~"), "RSASEG.egp", sep = "\\\\")
dbDisconnect(conn, projectPath = RSASEG_project)
}
}
\seealso{
\code{\link[=dbDisconnect,SASEGConnection-method]{dbDisconnect}}
}
