% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\docType{class}
\name{SASEGApplication-class}
\alias{SASEGApplication-class}
\alias{SASEGApplication}
\alias{show,SASEGApplication-method}
\alias{setProfile,SASEGApplication-method}
\alias{terminate,SASEGApplication-method}
\alias{newProject,SASEGApplication-method}
\title{An S4 class to represent a SAS EG Application object}
\usage{
SASEGApplication()

\S4method{show}{SASEGApplication}(object)

\S4method{setProfile}{SASEGApplication}(application, profile)

\S4method{terminate}{SASEGApplication}(application)

\S4method{newProject}{SASEGApplication}(application)
}
\arguments{
\item{application}{A \code{SASEGApplication} object.}

\item{profile}{A character string with the SAS profile name.}
}
\value{
A \code{SASEGApplication} object.
}
\description{
The \code{SASEGApplication} class is an S4 class to represent a \code{SAS EG Scripting Application} object.
}
\examples{
\dontrun{
# Modify the path below following your install:
path <- "C:/Program Files/SAS94/SASEnterpriseGuide/7.1/SASEGScripting.dll"
loadSASEGScripting(path)

app <- SASEGApplication()
show(app)

my_profile <- "PROFILE"
setProfile(app, my_profile)

my_new_project <- newProject(app)

# Important: you have to quit SAS EG Application
terminate(app)
}
}
\seealso{
\code{\linkS4class{SASEGProject}}
}
