% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\docType{class}
\name{SASEGApplication-class}
\alias{SASEGApplication-class}
\alias{SASEGApplication}
\alias{getName,SASEGApplication-method}
\alias{getVersion,SASEGApplication-method}
\alias{show,SASEGApplication-method}
\alias{setProfile,SASEGApplication-method}
\alias{terminate,SASEGApplication-method}
\alias{newProject,SASEGApplication-method}
\title{An S4 class to represent a SAS EG Application object}
\usage{
SASEGApplication()

\S4method{getName}{SASEGApplication}(object, ...)

\S4method{getVersion}{SASEGApplication}(object, ...)

\S4method{show}{SASEGApplication}(object)

\S4method{setProfile}{SASEGApplication}(application, profile)

\S4method{terminate}{SASEGApplication}(application)

\S4method{newProject}{SASEGApplication}(application)
}
\arguments{
\item{object}{An object created by \code{SASEGApplication()}.}

\item{application}{A \code{SASEGApplication} object.}

\item{profile}{A character string with the \code{SAS EG} profile name.}
}
\value{
\code{SASEGApplication()} is the constructor; it returns a 
    \code{SASEGApplication} object.

\code{terminate} method returns \code{TRUE}.

\code{newProject} returns a \code{\linkS4class{SASEGProject}} object.
}
\description{
The \code{SASEGApplication} class is an S4 class to represent a 
\code{SAS EG Scripting Application} object.

\code{getName} returns the name of a \code{SASEGApplication} object.

\code{getVersion} returns the version number of a \code{SASEGApplication} object.

\code{show} method provides useful informations including 
    \code{SAS EG} version and available profiles.

\code{setProfile} method is used to set the profile in a \code{SASEGApplication} object.

\code{terminate} method is used to quit a \code{SASEGApplication}.

\code{newProject} method opens a new \code{SASEGProject} in a given 
    \code{SASEGApplication}.
}
\examples{
\dontrun{
# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
loadSASEGScripting(path)

app <- SASEGApplication()
show(app)

my_profile <- "PROFILE"
setProfile(app, my_profile)

my_new_project <- newProject(app)

# Important: you have to quit SAS EG Application
terminate(app)
}
}
\seealso{
\code{\linkS4class{SASEGProject}}
}
