% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\docType{class}
\name{SASEGCode-class}
\alias{SASEGCode-class}
\alias{setServer,SASEGCode-method}
\alias{setText,SASEGCode-method}
\alias{setName,SASEGCode-method}
\alias{run,SASEGCode-method}
\alias{getLog,SASEGCode-method}
\alias{countOutputDatasets,SASEGCode-method}
\alias{getSourceCode,SASEGCode-method}
\alias{getListDatasets,SASEGCode-method}
\title{An S4 class to represent a SAS EG Code object}
\usage{
\S4method{setServer}{SASEGCode}(object, server, ...)

\S4method{setText}{SASEGCode}(object, text, ...)

\S4method{setName}{SASEGCode}(object, name, ...)

\S4method{run}{SASEGCode}(code)

\S4method{getLog}{SASEGCode}(code)

\S4method{countOutputDatasets}{SASEGCode}(code)

\S4method{getSourceCode}{SASEGCode}(code)

\S4method{getListDatasets}{SASEGCode}(code)
}
\arguments{
\item{object}{A \code{SASEGCode} object.}

\item{server}{A character string with the \code{SAS} server name.}

\item{...}{Other parameters. Not used.}

\item{text}{A character string with the code of the \code{SAS} program.}

\item{name}{A character string with the name of the program.}

\item{code}{A \code{SASEGCode} object.}
}
\value{
\code{getLog} returns a character string containing the log.

\code{countOutputDatasets} returns a numeric.

\code{getSourceCode} returns a character string.

\code{getListDatasets} returns a list of \code{\linkS4class{SASEGDataset}} 
    objects named with their filenames (ie. "LIBNAME.NAME").
}
\description{
The \code{SASEGCode} class is an S4 class to represent a 
\code{SAS EG Scripting Code} object.

\code{setServer} method is used to set the \code{SAS} server that will be 
   used to execute the program contained in the \code{SASEGCode} object.

\code{setText} method is used to put a \code{SAS} program in a \code{SASEGCode} object.

\code{setName} method is used to give a name to a \code{SASEGCode} object.

\code{run} method is used to run (or execute) the \code{SAS} program of a 
    \code{SASEGCode} object.

\code{getLog} method is used to retrieve the log after an execution.

\code{countOutputDatasets} method is used to get the number of datasets 
    created by the code execution.

\code{getSourceCode} method is used to retrieve the source code that was executed.

\code{getListsDatasets} method is used to retrieve a named list of 
    \code{SASEGDataset} objects created by the code execution.
}
\section{Constructor}{

There is no constructor. To create a new \code{SASEGCode} object,
you have to call the \code{\linkS4class{SASEGProject}} method 
\code{\link{newCode}}.
}

\examples{
\dontrun{
# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
loadSASEGScripting(path)
app <- SASEGApplication()
setProfile(app, "PROFILE")
my_new_project <- newProject(app)
my_new_code <- newCode(my_new_project)

setServer(my_new_code, "SASPROD")
setText(my_new_code, "data a; set sashelp.class; run;")
setName(my_new_code, "Test RSASEG")
run(my_new_code)
log <- getLog(my_new_code)
cat(log)
n <- countOutputDatasets(my_new_code)
check_my_code <- getSourceCode(my_new_code)
cat(check_my_code)
a <- getListDatasets(my_new_code)

# Important: you have to quit SAS EG Application
terminate(app)
}
}
\seealso{
\code{\linkS4class{SASEGProject}}, \code{\linkS4class{SASEGDataset}}
}
