% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\docType{class}
\name{SASEGDataset-class}
\alias{SASEGDataset-class}
\alias{getFileName,SASEGDataset-method}
\alias{getName,SASEGDataset-method}
\alias{saveAs,SASEGDataset-method}
\alias{read,SASEGDataset-method}
\title{An S4 class to represent a SAS EG OutputDataset object}
\usage{
\S4method{getFileName}{SASEGDataset}(object)

\S4method{getName}{SASEGDataset}(object)

\S4method{saveAs}{SASEGDataset}(object, dir = NULL, name = NULL,
  type = "tab", fsep = "\\\\")

\S4method{read}{SASEGDataset}(object)
}
\arguments{
\item{object}{A \code{SASEGDataset} object.}

\item{dir}{Optional, a character string with the path of a directory. If \code{NULL}, the 
temporary directory of the session is used.}

\item{name}{Optional, a character string with the name of the file (without extension). If 
\code{NULL}, the name of the dataset is used.}

\item{type}{Optional, a character with the file format (\code{"csv"}, 
\code{"tab"} or \code{"txt"}). \code{tab} is used by default.}

\item{fsep}{Optional, a character with a path separator passed to 
\code{\link[base]{file.path}}. As \code{SAS EG} is a \code{Windows} 
application, "\\" is used by default.}

\item{object}{A \code{SASEGDataset} object.}
}
\value{
\code{getFileName} returns a character string with the filename of the dataset 
    (ie. \emph{with} libname).

\code{getName} returns a character string with the name of the dataset 
    (ie. \emph{without} libname).

\code{saveAs} method returns a character string containing the filepath.

\code{read} returns a \code{data.table}.
}
\description{
The \code{SASEGDataset} object is an S4 class to represent a \code{SAS EG
Scripting OutputDataset} object.

\code{getFileName} method is used to get the filename of a \code{SASEGDataset}.

\code{getName} method is used to get the name of a \code{SAS} dataset.

\code{saveAs} method is used to save a dataset in a local file.

\code{read} method is used to read a \code{SASEGDataset} and retrieve data 
    in a \code{data.table}.
}
\details{
\code{read} method stores data in a local temporary file and then imports 
    in \code{R} with \code{\link[data.table]{fread}}.
}
\section{Constructor}{
 There is no constructor. \code{SASEGDataset} objects
  result from a \code{\linkS4class{SASEGCode}} execution, after calling the 
  \code{\linkS4class{SASEGCode}} method \code{\link{getListDatasets}}.
}

\examples{
\dontrun{
# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
loadSASEGScripting(path)
app <- SASEGApplication()
setProfile(app, "PROFILE")
my_new_project <- newProject(app)
my_server <- "SASPROD"
sas_program <- "data a; set sashelp.class; run;"
program_name <- "Test RSASEG"
my_new_code <- newCode(my_new_project,
                       my_server,
                       sas_program,
                       program_name)
run(my_new_code)
a <- getListDatasets(my_new_code)

getName(a[[1]])
getFileName(a[[1]])
path_to_csv <- saveAs(a[[1]], normalizePath("~"))
a.df <- data.table::fread(path_to_csv)                           

# Important: you have to quit SAS EG Application
terminate(app)
}
}
\seealso{
\code{\linkS4class{SASEGCode}}
}
