% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGDriver-class}
\alias{SASEGDriver-class}
\alias{app,SASEGDriver-method}
\alias{cnx<-,SASEGDriver-method}
\alias{isValid<-,SASEGDriver-method}
\alias{dbGetInfo,SASEGDriver-method}
\title{Driver class for SAS Enterprise Guide}
\usage{
\S4method{app}{SASEGDriver}(obj, ...)

\S4method{cnx}{SASEGDriver}(obj) <- value

\S4method{isValid}{SASEGDriver}(obj) <- value

\S4method{dbGetInfo}{SASEGDriver}(dbObj, ...)
}
\arguments{
\item{obj}{A \code{SASEGDriver} object.}

\item{...}{Other parameters passed on to method. Not used.}

\item{value}{A value to replace for.}

\item{dbObj}{A \code{SASEGDriver} object.}
}
\description{
This a driver class for \code{SAS Enterprise Guide}. The \code{SASEGDriver} 
    class inherits from the \code{\link[DBI]{DBIDriver-class}}. A 
    \code{SASEGDriver} object can be understood as a \code{SAS EG} application.
    \code{SASEGDriver} objects are constructed by calling 
    \code{\link[=SASEG]{SASEG()}}.

\code{app()} access to the slot \code{app} of a \code{SASEGDriver} object. 
    This method is not exported. Only developpers may need to use it.

\code{`cnx<-`} replaces value stored in slot \code{cnx}. If there is already an 
    active connection, the only value accepted is \code{NULL}, an error is 
    raised otherwise. If there is no active connection, an object of class 
    \code{\linkS4class{SASEGConnection}} is accepted as a value. This method 
    is not exported. Only developpers may need to use it.

Value stored in slot \code{isValid} can be replaced with \code{`isValid<-`}. This 
    method is not exported. Only developpers may need to use it.

\code{dbGetInfo} method is called to get:
\itemize{
    \item \code{driver.version}: version of \code{RSASEG} package.
    \item \code{client.version}: version of \code{SAS Enterprise Guide}.
    \item \code{max.connections}: equals to 1.
    \item \code{available.profiles}: a list with informations about 
        available profiles.
}
}
\section{Slots}{

\describe{
\item{\code{app}}{A \code{\linkS4class{SASEGApplication}} object.}

\item{\code{cnx}}{A closure in which active connections are stored. Values can be 
accessed with 
\code{\link[=dbListConnections,SASEGDriver-method]{dbListConnections}}.}

\item{\code{isValid}}{A closure in which the status of the object is stored. Value 
can be accessed with 
\code{\link[=dbIsValid,SASEGDriver-method]{dbIsValid}}.}
}}

\examples{
\dontrun{
library(DBI)
library(rClr)

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
drv <- RSASEG::SASEG(path)

dbGetInfo(drv)}
}
\keyword{internal}
