% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGDriver-class}
\alias{SASEGDriver-class}
\alias{app,SASEGDriver-method}
\title{Driver class for SAS Enterprise Guide}
\usage{
\S4method{app}{SASEGDriver}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{SASEGDriver}}.}

\item{...}{Other parameters passed to method. Not used.}
}
\description{
This a driver class for \code{SAS Enterprise Guide}. The \code{SASEGDriver} 
    class inherits from the \code{\link[DBI]{DBIDriver-class}}. A 
    \code{SASEGDriver} object can be understood as a \code{SAS EG} application.

\code{app()} access to the slot \code{app} of a \code{\linkS4class{SASEGDriver}}.
}
\keyword{internal}
