% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGDriver-class}
\alias{SASEGDriver-class}
\alias{app,SASEGDriver-method}
\alias{cnx<-,SASEGDriver-method}
\alias{isValid<-,SASEGDriver-method}
\title{Driver class for SAS Enterprise Guide}
\usage{
\S4method{app}{SASEGDriver}(obj, ...)

\S4method{cnx}{SASEGDriver}(obj) <- value

\S4method{isValid}{SASEGDriver}(obj) <- value
}
\arguments{
\item{obj}{A \code{SASEGDriver} object.}

\item{...}{Other parameters passed on to method. Not used.}

\item{value}{A value to replace for.}
}
\description{
This a driver class for \code{SAS Enterprise Guide}. The \code{SASEGDriver} 
    class inherits from the \code{\link[DBI]{DBIDriver-class}}. A 
    \code{SASEGDriver} object can be understood as a \code{SAS EG} application.
    \code{SASEGDriver} objects are constructed by calling 
    \code{\link[=SASEG]{SASEG()}}.

\code{app()} access to the slot \code{app} of a \code{SASEGDriver} object. 
    This method is not exported. Only developpers may need to use it.

\code{`cnx<-`} replaces value stored in slot \code{cnx}. If there is already an 
    active connection, the only value accepted is \code{NULL}, an error is 
    raised otherwise. If there is no active connection, an object of class 
    \code{\linkS4class{SASEGConnection}} is accepted as a value. This method 
    is not exported. Only developpers may need to use it.

Value stored in slot \code{isValid} can be replaced with \code{`isValid<-`}. This 
    method is not exported. Only developpers may need to use it.
}
\section{Slots}{

\describe{
\item{\code{app}}{A \code{\linkS4class{SASEGApplication}} object.}

\item{\code{cnx}}{A closure in which active connections are stored. Values can be 
accessed with 
\code{\link[=dbListConnections,SASEGDriver-method]{dbListConnections}}.}

\item{\code{isValid}}{A closure in which the status of the object is stored. Value 
can be accessed with 
\code{\link[=dbIsValid,SASEGDriver-method]{dbIsValid}}.}
}}

\seealso{
\code{SASEGDriver} methods: 
    \code{\link[=dbGetInfo,SASEGDriver-method]{dbGetInfo}},
    \code{\link[=dbIsValid,SASEGDriver-method]{dbIsValid}},
    \code{\link[=dbListConnections,SASEGDriver-method]{dbListConnections}},
    \code{\link[=dbUnloadDriver,SASEGDriver-method]{dbUnloadDriver}},
    \code{\link[=dbDataType,SASEGDriver-method]{dbDataType}}
}
\keyword{internal}
