% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\docType{class}
\name{SASEGProject-class}
\alias{SASEGProject-class}
\alias{saveAs,SASEGProject-method}
\alias{terminate,SASEGProject-method}
\alias{newCode,SASEGProject-method}
\title{An S4 class to represent a SAS EG Project object}
\usage{
\S4method{saveAs}{SASEGProject}(object, filepath, ...)

\S4method{terminate}{SASEGProject}(object)

\S4method{newCode}{SASEGProject}(project, server = NULL, program = NULL,
  name = NULL, ...)
}
\arguments{
\item{object}{A \code{SASEGProject} object.}

\item{filepath}{A character string with a valid path.
\strong{Be careful: \code{saveAs} method overwrites existing files 
without confirmation.}}

\item{...}{Other parameters passed on to method (not used).}

\item{project}{A \code{SASEGProject} object.}

\item{server}{A character string with the server name. Optional, 
\code{NULL} by default.}

\item{program}{A character string with the \code{SAS} code of the program.
Optional, \code{NULL} by default.}

\item{name}{A character string with the name of the program. Optional, 
\code{NULL} by default.}

\item{object}{A \code{SASEGProject} object.}
}
\value{
\code{newCode} method returns a \code{\linkS4class{SASEGCode}}.
}
\description{
The \code{SASEGProject} class is an S4 class to represent a 
    \code{SAS EG Scripting Project} object.

\code{saveAs} method saves a \code{SASEGProject} to a file.

\code{terminate} method closes a \code{SASEGProject}.

\code{newCode} method is used to create a new \code{SASEGCode} object in a 
    \code{SASEGProject} object.
}
\details{
\code{saveAs} method shows a message after saving the file.
}
\section{Constructor}{

There is no constructor. To create a new \code{SASEGProject} object, you have 
    to call the \code{\linkS4class{SASEGApplication}} method 
    \code{\link{newProject}}.
}

\examples{
\dontrun{
# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"
loadSASEGScripting(path)
app <- SASEGApplication()

my_profile <- "PROFILE"
setProfile(app, my_profile)

my_new_project <- newProject(app)

my_server <- "SASPROD"
sas_program <- "data a; set sashelp.class; run;"
program_name <- "Test RSASEG"
my_new_code <- newCode(my_new_project,
                       my_server,
                       sas_program,
                       program_name)
                       
path <- paste(normalizePath("~"), "RSASEG.egp", sep = "\\\\")
saveAs(my_new_project, path)

# Important: you have to quit SAS EG Application
terminate(app)
}
}
\seealso{
\code{\linkS4class{SASEGApplication}}, \code{\linkS4class{SASEGCode}}
}
\keyword{internal}
