% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGResult-class}
\alias{SASEGResult-class}
\title{SASEG results class}
\description{
\code{SASEGResult} class inherits from \code{\link[DBI]{DBIResult-class}}.
}
\section{Slots}{

\describe{
\item{\code{SASResult}}{An object of class \code{\linkS4class{SASEGCode}}.}

\item{\code{SASUtil}}{An object of class \code{\linkS4class{SASEGCode}}.}

\item{\code{SQLResult}}{A character string. This slot contains the filename of the 
dataset created by the \code{ODS} during a \code{PROC SQL}.}

\item{\code{fetch}}{A closure.}

\item{\code{rowsFetched}}{A closure.}

\item{\code{isValid}}{A closure.}
}}

\keyword{internal}
