% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSASEGDriver.R
\docType{class}
\name{SASEGResult-class}
\alias{SASEGResult-class}
\title{SASEG results class}
\description{
\code{SASEGResult} class inherits from \code{\link[DBI]{DBIResult-class}}. 
   This class represents results of \emph{any} \code{SAS} program. It is 
   extended by \code{\linkS4class{SASEGSQLResult}} for \code{SQL} results.
}
\section{Slots}{

\describe{
\item{\code{conn}}{An object of class \code{\linkS4class{SASEGConnection}}.}

\item{\code{SASResult}}{An object of class \code{\linkS4class{SASEGCode}}.}

\item{\code{fetched}}{A closure.}

\item{\code{isValid}}{A closure.}
}}

\keyword{internal}
