% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\docType{class}
\name{SASObjRef-class}
\alias{SASObjRef-class}
\alias{SASObjRef}
\alias{ptr,SASObjRef-method}
\alias{clrCall,SASObjRef-method}
\alias{clrGet,SASObjRef-method}
\alias{clrSet,SASObjRef-method}
\title{An S4 virtual class to represent any SAS EG Scripting object}
\usage{
SASObjRef(ptr)

\S4method{ptr}{SASObjRef}(SASObj)

\S4method{clrCall}{SASObjRef}(obj, methodName, ...)

\S4method{clrGet}{SASObjRef}(objOrType, name)

\S4method{clrSet}{SASObjRef}(objOrType, name, value)
}
\arguments{
\item{ptr}{A \code{CLR} object. Any other \code{R} object can be passed on argument.}

\item{SASObj}{A \code{SASObjRef} object.}

\item{obj}{A \code{SASObjRef} object.}

\item{methodName}{A character string with the name of a method of the object.}

\item{...}{Additional method arguments.}

\item{objOrType}{A \code{SASObjRef} object.}

\item{name}{The name of a field/property of the object.}

\item{value}{The value to set the field with.}
}
\value{
\code{SASObjRef(ptr)} returns a \code{SASObjRef} object if \code{ptr} is a 
    \code{CLR} object; returns \code{ptr} otherwise.

\code{ptr(SASObj)} returns a \code{cobjRef} object from a \code{SASObjRef} object.

\code{clrCall} returns an object from the call. May be a 
    \code{SASObjRef} object, or a native \code{R} object for common types.
    Can be \code{NULL}.

\code{clrGet} returns an object from the call. May be a 
    \code{SASObjRef} object, or a native \code{R} object for common types. 
    Can be \code{NULL}.
}
\description{
The \code{SASObjRef} S4 class represents any \code{SAS EG Scripting} object.
This is a virtual class.
This class and its methods are internals.
}
\details{
A \code{SASObjRef} object has only one slot, named \code{ptr}, where \code{ptr} is
a \code{cobjRef} object. For more information about \code{CLR} objects and methods,
see the \code{rClr} package documentation.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{A \code{cobjRef} object.}
}}

\examples{
\dontrun{

# Modify the path below following your install:
path <- "C:\\\\Program Files\\\\SAS94\\\\SASEnterpriseGuide\\\\7.1\\\\SASEGScripting.dll"

loadSASEGScripting(path)

app_clrobj <- rClr::clrNew("SAS.EG.Scripting.Application")
app_sasobj <- SASObjRef(app_clrobj)

show(app_clrobj)
show(ptr(app_sasobj))

clrGet(app_sasobj, "Name")

clrCall(app_sasobj, "Quit")
}
}
\seealso{
Package \code{rClr}: \code{\link[rClr]{clrCall}}, \code{\link[rClr]{clrGet}}, 
    \code{\link[rClr]{clrSet}}

Generics:  \code{\link{clrCall}}, \code{\link{clrGet}}, \code{\link{clrSet}}
}
\keyword{internal}
