% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASEGS4.R
\name{terminate}
\alias{terminate}
\title{Terminate (or quit) an application}
\usage{
terminate(application)
}
\arguments{
\item{application}{An application object.}
}
\description{
\code{terminate} method is used to quit an application.
}
\keyword{internal}
